Question, Answer 엔티티에 author 필드 추가

+++ src/test/java/com/exam/sbb/QuestionRepositoryTests.java
@@ -21,7 +21,7 @@ class QuestionRepositoryTests {
 
 	@Autowired
 	private QuestionRepository questionRepository;
-	private static Long lastSampleDataId;
+	private static int lastSampleDataId;
 
 	@BeforeEach
 	void beforeEach() {
@@ -29,7 +29,7 @@ class QuestionRepositoryTests {
 		createSampleData();
 	}
 
-	public static Long createSampleData(QuestionRepository questionRepository) {
+	public static int createSampleData(QuestionRepository questionRepository) {
 		Question q1 = new Question();
 		q1.setSubject("sbb가 무엇인가요?");
 		q1.setContent("sbb에 대해서 알고 싶습니다.");
@@ -148,7 +148,7 @@ class QuestionRepositoryTests {
 	@Test
 	void Pageable() {
 		// Pageable : 한 페지이에 몇개의 아이템이 나와야 하는지 + 현재 몇 페이지인지
-		Pageable pageable = PageRequest.of(0, Math.toIntExact(lastSampleDataId));
+		Pageable pageable = PageRequest.of(0, lastSampleDataId);
 
 		Page<Question> page = questionRepository.findAll(pageable);
